Remove .travis.yml
The fmt and validate commands for Terraform 0.12 require an initialized
directory. For this module that would involve providing read access to
Travis for the S3 state bucket. Given the amount of sensitive
information in the state files I'm hesitant to do so. fmt should be
configured to run automatically on save in your editor anyways.