Add shared logging outputs from global remote state
Why these changes are being introduced:
The resources deployed in TF_WORKSPACE=global have been updated and
there are three new outputs from the code in mitlib-terraform/global
that need to be made available to apps in mitlib-terraform/apps.

How this addresses that need:
* Three outputs are added to outputs.tf

Side effects of this change:
This will force us to remove and re-attach the "0.13" tagged release
so that these outputs are available to apps without having to update
the module call for those apps.

Relevant PR/merge:
* https://github.com/MITLibraries/mitlib-terraform/pull/265